provide a way for non-vcl modules to register factories
There are still two huge problems around my approach. It still requires
to add some code to the existing widget class so that it registers the
UI object wrapper. And secondly I have not yet solved the problem of
longer inheritance hierarchies. E.g. A <- B <- C.

However it is a first version that allows for now to extend the UI
testing code into other places. I'm soon moving the header for the
ui objects to the public include so that all objects can inherit from
the basic vcl wrapper objects.

Change-Id: Ifac9d635b4a68737048f50cd2a9d3e8830b165a5